requirements/develop.txt
amqp==2.3.2
azure==0.11.1
billiard==3.5.0.4
cachetools==2.0.1
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
diff-match-patch==20121119
Django==1.11.15
django-cors-headers==2.3.0
django-csvexport==1.1.5
django-import-export==1.0.1
django-multiselectfield==0.1.8
django-storages==1.6.6
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-yasg==1.10.2
et-xmlfile==1.0.1
flex==6.13.2
future==0.15.2
google-api-core==1.1.1
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
inflection==0.3.1
ipython==6.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4
jedi==0.11.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
kombu==4.2.1
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
olefile==0.44
openpay==0.6
openpyxl==2.5.4
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pudb==2017.1.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
PyJWT==1.6.1
python-dateutil==2.6.1
python-http-client==3.1.0
pytz==2017.2
PyYAML==3.13
raven==6.10.0
redis==2.10.6
requests==2.18.4
rfc3987==1.3.8
rsa==3.4.2
ruamel.yaml==0.15.69
sendgrid==5.6.0
simplegeneric==0.8.1
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.0
tablib==0.12.1
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
Unipath==1.1
uritemplate==3.0.0
urllib3==1.22
urwid==1.3.1
validate-email==1.3
vine==1.1.4
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
